fix(doctor): use u8 for exit codes to avoid lossy cast
Address review comment about dangerous i32 to u8 cast when converting
doctor exit codes to ExitCode type.

Changes:
- Change DoctorReport::exit_code() return type from i32 to u8
- Update doctor() function signature to return Result<u8>
- Update output_json() function signature to return Result<u8>
- Remove lossy 'as u8' cast in main.rs doctor command dispatch

The exit codes remain semantically identical (0=pass, 1=warn, 2=fail)
but now use the correct type throughout the call chain.

+++ crates/claudio/src/commands/doctor.rs
@@ -57,7 +57,7 @@ impl DoctorReport {
         self.checks.push(result);
     }
 
-    fn exit_code(&self) -> i32 {
+    fn exit_code(&self) -> u8 {
         if self.summary.fail > 0 {
             2
         } else if self.summary.warn > 0 {
@@ -73,7 +73,7 @@ pub fn doctor(
     json_output: bool,
     verbose: bool,
     claude_executable_cli: Option<&str>,
-) -> Result<i32> {
+) -> Result<u8> {
     let effective_scope = effective_read_scope(scope);
 
     let claude_executable =
@@ -492,7 +492,7 @@ fn output_human_readable(report: &DoctorReport) {
     println!();
 }
 
-fn output_json(report: &DoctorReport) -> Result<i32> {
+fn output_json(report: &DoctorReport) -> Result<u8> {
     let output = json!({
         "checks": report.checks,
         "summary": {
